Runtime/RTAcpi: Some more work on the ASL -> AML compiler, can process our vbox-standard.dsl now, bugref:10733 [build fix]
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108030 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/acpi/acpi-compiler.cpp
@@ -699,14 +699,14 @@ static int rtAcpiTblAslParseTerminal(PRTACPIASLCU pThis, RTACPIASLTERMINAL enmTe
         RTACPIASL_PARSE_PUNCTUATOR(RTACPIASLTERMINAL_PUNCTUATOR_OPEN_BRACKET, '(');
 
         /* Process any required arguments. */
-        for (uint32_t i = 0; i < pAslKeyword->cArgsReq; i++)
+        for (uint8_t i = 0; i < pAslKeyword->cArgsReq; i++)
         {
             rc = rtAcpiTblAslParseArgument(pThis, pAslKeyword->pszOpc, i, pAslKeyword->aenmTypes[i], &pAstNd->aArgs[i]);
             if (RT_FAILURE(rc))
                 return rc;
 
             /* There must be a "," between required arguments, not counting the last required argument because it can be closed with ")". */
-            if (i < pAslKeyword->cArgsReq - 1)
+            if (i < (uint8_t)(pAslKeyword->cArgsReq - 1))
                 RTACPIASL_PARSE_PUNCTUATOR(RTACPIASLTERMINAL_PUNCTUATOR_COMMA, ',');
         }
